+++ telefunc/babel.js
@@ -1,2 +1,2 @@
 // React Native Expo/Metro doesn't seem to support `package.json#exports`.
-module.exports = require('./dist/node/babel/babel-plugin-telefunc.js')
+module.exports = require('./dist/node/babel/plugin.js')
